Cleaned up main-page-update
