+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "prepublish": "BUILD_DEV=0 API_ENV=production SEGMENT_API_ENV=production webpack && cp src/_redirects public/_redirects",
-    "_stage": "surge -p public -d test.getguesstimate.com",
+    "_stage": "surge -p public -d stage.getguesstimate.com",
     "start": "BUILD_DEV=1 NODE_ENV=development API_ENV=development SEGMENT_API_ENV=development webpack-dev-server",
     "start-production": "BUILD_DEV=1 NODE_ENV=development API_ENV=production SEGMENT_API_ENV=development webpack-dev-server",
     "stage": "npm run prepublish && npm run _stage",
