+++ src/components/spaces/cards/index.js
@@ -1,37 +1,13 @@
 import React from 'react'
 import Icon from'react-fa'
+
+import arrowsVisibleImage from '../../../assets/metric-icons/blue/arrows-visible.png'
+import {formatDescription, formatDate} from 'gComponents/spaces/shared'
 import './style.css'
-import moment from 'moment'
-import removeMd from 'remove-markdown'
+
 import * as navigationActions from 'gModules/navigation/actions.js'
 import * as Space from 'gEngine/space';
 import * as User from 'gEngine/user';
-import arrowsVisibleImage from '../../../assets/metric-icons/blue/arrows-visible.png'
-
-function formatDescription(description) {
-  const maxLength = 300
-
-  if (_.isEmpty(description)){ return '' }
-
-  const withoutMarkdown = removeMd(description)
-  if (withoutMarkdown.length < maxLength) { return withoutMarkdown }
-
-  const truncated = withoutMarkdown.substring(0, maxLength)
-  return `${truncated}...`
-}
-
-function formatDate(date) {
- return moment(new Date(date)).format('ll')
-}
-
-let PrivateTag = ({isPrivate}) => (
-  <div className='col-xs-12'>
-    <div className='privacy-tag'>
-      <Icon name={isPrivate ? 'lock' : 'globe'}/>
-      {isPrivate ? 'Private' : 'Public'}
-    </div>
-  </div>
-)
 
 let BlankScreenshot = () => (
   <div className='snapshot blank'>
@@ -39,7 +15,7 @@ let BlankScreenshot = () => (
   </div>
 )
 
-let MinorButton = ({isPrivate}) => (
+let SingleButton = ({isPrivate}) => (
   <div className='tag'>
     <Icon name={isPrivate ? 'lock' : 'globe'}/>
   </div>
@@ -58,7 +34,7 @@ let ButtonArea = ({user, navigateToUser, isPrivate, showPrivacy}) => (
       </div>
     </div>
     }
-    {showPrivacy && <MinorButton isPrivate={isPrivate}/>}
+    {showPrivacy && <SingleButton isPrivate={isPrivate}/>}
   </div>
 )
 
@@ -67,8 +43,8 @@ const SpaceCard = ({space, showPrivacy}) => {
   const navigateToSpace = () => {navigationActions.navigate(Space.url(space))}
   const navigateToUser = () => {navigationActions.navigate(User.url(space.user))}
   return (
-    <div className='col-md-4 SpaceCard-same-height'>
-      <div className='SpaceCard'>
+    <div className='col-md-4 SpaceCard'>
+      <div className='SpaceCard--inner'>
         <div className={`header ${hasName ? '' : 'default-name'}`}>
           <h3 onClick={navigateToSpace}>{hasName ? space.name : 'Untitled Model'}</h3>
           <div className='changed-at'>Updated {formatDate(space.updated_at)}</div>
@@ -92,22 +68,15 @@ const SpaceCard = ({space, showPrivacy}) => {
   )
 }
 
-const SpaceCards = ({spaces, hasMorePages, loadMore, showPrivacy}) => (
-  <div className='Cards row'>
-    {_.map(spaces, (s) => {
-      return (
+const SpaceCards = ({spaces, showPrivacy}) => (
+  <div className='row'>
+    {_.map(spaces, (s) =>
         <SpaceCard
           key={s.id}
           space={s}
           showPrivacy={showPrivacy}
         />
-      )
-    })}
-    {!!spaces.length && hasMorePages &&
-      <div className='nextPage'>
-        <button className={'ui button nextpage large'} onClick={loadMore}> {'Load More'} </button>
-      </div>
-    }
+    )}
   </div>
 )
