+++ src/components/spaces/index/index.js
@@ -3,12 +3,12 @@ import Icon from'react-fa'
 import { connect } from 'react-redux';
 import SpaceCards from 'gComponents/spaces/cards'
 import * as search from 'gModules/search_spaces/actions'
+import './style.css'
 
 function mapStateToProps(state) {
   return {
     searchSpaces: state.searchSpaces,
     users: state.users,
-    me: state.me
   }
 }
 
@@ -28,20 +28,21 @@ export default class SpacesIndex extends Component{
     const {searchSpaces, showScreenshots} = this.props
     let spaces = searchSpaces.hits || []
     const hasMorePages = _.isFinite(searchSpaces.page) && (searchSpaces.page < (searchSpaces.nbPages - 1))
-    const size='SMALL'
     return (
-      <div>
+      <div className='SpacesIndex'>
         <div className='row'>
           <div className='col-md-4'/>
           <div className='col-md-4'>
-            <div className='stuff search-form'>
+            <div className='search-form'>
               <div className='ui form'>
                 <input name='search' placeholder='Search Models' onChange={this._search.bind(this)}/>
               </div>
             </div>
           </div>
         </div>
-        <SpaceCards showPrivacy={false} spaces={spaces.map(s => {return {...s, user: s.user_info}})} size={size}/>
+
+        <SpaceCards showPrivacy={false} spaces={spaces.map(s => {return {...s, user: s.user_info}})}/>
+
         {!!spaces.length && hasMorePages &&
           <div className='nextPage'>
             <button className={'ui button nextpage large'} onClick={this._nextPage.bind(this)}> {'Load More'} </button>
