+++ src/components/spaces/index/style.css
@@ -1,0 +1,11 @@
+.SpacesIndex .search-form {
+  margin-top:3em;
+  margin-bottom: 3em;
+}
+
+.SpacesIndex .nextPage {
+  margin-bottom: 4em;
+  margin-top: 2em;
+  text-align: center;
+}
+
