+++ src/components/spaces/shared.js
@@ -1,0 +1,19 @@
+import removeMd from 'remove-markdown'
+import moment from 'moment'
+
+export function formatDescription(description) {
+  const maxLength = 300
+
+  if (_.isEmpty(description)){ return '' }
+
+  const withoutMarkdown = removeMd(description)
+  if (withoutMarkdown.length < maxLength) { return withoutMarkdown }
+
+  const truncated = withoutMarkdown.substring(0, maxLength)
+  return `${truncated}...`
+}
+
+export function formatDate(date) {
+ return moment(new Date(date)).format('ll')
+}
+
