--- src/components/spaces/shared/style.css
@@ -1,22 +1,0 @@
-@import './variables.css';
-
-.GeneralSpaceIndex {
-  float: left;
-  width: 100%;
-}
-
-.GeneralSpaceIndex  h2 {
-  margin-top: 10px;
-  color: $blue-1;
-  font-family: Lato;
-  margin-bottom: 0.5em;
-  text-align: center;
-}
-
-.GeneralSpaceIndex > .divider.ui.divider {
-  margin-top: 2em;
-  margin-bottom: 3em;
-  border-color: $grey-1;
-  border-bottom: none;
-}
-
