+++ src/components/users/show/style.css
@@ -1,15 +1,15 @@
-.userShow .main-user-tag {
+.UserShow .main-user-tag {
   text-align: center;
   margin-bottom: 2em;
 }
 
-.userShow .main-user-tag img{
+.UserShow .main-user-tag img{
   width: 8em;
   height: 8em;
   border-radius: 8em;
 }
 
-.userShow .main-user-tag h1{
+.UserShow .main-user-tag h1{
   margin-bottom: 0;
   margin-top: 0.5rem;
 }
